+++ depthai_ros_driver/include/depthai_ros_driver/param_handlers/nn_param_handler.hpp
@@ -43,14 +43,7 @@ class NNParamHandler : public BaseParamHandler {
         declareAndLogParam<bool>("i_enable_passthrough", false);
         declareAndLogParam<bool>("i_enable_passthrough_depth", false);
         declareAndLogParam<bool>("i_get_base_device_timestamp", false);
-        dai::CameraBoardSocket socket = dai::CameraBoardSocket::RGB;
-        try {
-            socket = static_cast<dai::CameraBoardSocket>(getROSNode()->get_parameter("rgb.i_board_socket_id").as_int());
-        } catch(rclcpp::exceptions::ParameterNotDeclaredException& e) {
-            RCLCPP_ERROR(getROSNode()->get_logger(), "Unable to get correct socket for RGB node, defaulting to RGB.");
-        }
-        declareAndLogParam<int>("i_board_socket_id", static_cast<int>(socket));
-
+        declareAndLogParam<bool>("i_update_ros_base_time_on_ros_msg", false);
         auto nn_path = getParam<std::string>("i_nn_config_path");
         using json = nlohmann::json;
         std::ifstream f(nn_path);